requirements.txt
scipy
torch
tqdm
h5py
numpy
pandas
PyYAML
tensorboardX
